+++ src/main/resources/db/create.sql
@@ -1,16 +1,16 @@
 CREATE TABLE `category` (
-                            `id` bigint NOT NULL AUTO_INCREMENT,
+                            `id` int NOT NULL AUTO_INCREMENT,
                             `category_en` varchar(255) NOT NULL,
                             `category_ru` varchar(255) NOT NULL,
                             PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8
 
 CREATE TABLE `dish` (
-                        `id` bigint NOT NULL AUTO_INCREMENT,
+                        `id` int NOT NULL AUTO_INCREMENT,
                         `name_en` varchar(45) NOT NULL,
                         `name_ru` varchar(45) NOT NULL,
                         `price` decimal(9,2) NOT NULL,
-                        `category_id` bigint NOT NULL,
+                        `category_id` int NOT NULL,
                         `time` timestamp NOT NULL,
                         PRIMARY KEY (`id`),
                         KEY `fk_dish_category_idx` (`category_id`),
@@ -18,7 +18,7 @@ CREATE TABLE `dish` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8
 
 CREATE TABLE `login` (
-                         `id` bigint NOT NULL AUTO_INCREMENT,
+                         `id` int NOT NULL AUTO_INCREMENT,
                          `email` varchar(255) NOT NULL,
                          `login` varchar(255) NOT NULL,
                          `password` varchar(255) NOT NULL,
@@ -29,8 +29,8 @@ CREATE TABLE `login` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8
 
 CREATE TABLE `orders` (
-                          `id` bigint NOT NULL AUTO_INCREMENT,
-                          `login_id` bigint NOT NULL,
+                          `id` int NOT NULL AUTO_INCREMENT,
+                          `login_id` int NOT NULL,
                           `total_price` decimal(9,2) NOT NULL,
                           `status` varchar(45) NOT NULL,
                           `time` timestamp NOT NULL,
@@ -40,9 +40,9 @@ CREATE TABLE `orders` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8
 
 CREATE TABLE `cart` (
-                        `id` bigint NOT NULL AUTO_INCREMENT,
-                        `dish_id` bigint NOT NULL,
-                        `login_id` bigint NOT NULL,
+                        `id` int NOT NULL AUTO_INCREMENT,
+                        `dish_id` int NOT NULL,
+                        `login_id` int NOT NULL,
                         PRIMARY KEY (`id`),
                         KEY `fk_cart_to_dish_idx` (`dish_id`),
                         KEY `fk_cart_to_login_idx` (`login_id`),
